LayoutTests/platform/gtk/accessibility/custom-elements/modal-expected.txt

background

modal content
This tests that aria fallback roles work correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL !accessibilityController.accessibleElementById("background-content") || accessibilityController.accessibleElementById("background-content").isIgnored should be true. Was false.
PASS accessibilityController.accessibleElementById("background-content").isIgnored is false
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
